refactor: Change money currencies from int to decimal

+++ resources/init.sql
@@ -16,11 +16,11 @@ CREATE TABLE [User]
 
 CREATE TABLE [Bidder]
 (
-    [NIF]             INT   NOT NULL PRIMARY KEY,
-    [BirthDate]       DATE  NOT NULL,
-    [Balance]         INT   NOT NULL DEFAULT 0 CHECK ([Balance] >= 0),
-    [PendingBalance]  INT   NOT NULL DEFAULT 0 CHECK ([PendingBalance] >= 0),
-    [UserID]          INT   NOT NULL,
+    [NIF]            INT   NOT NULL PRIMARY KEY,
+    [BirthDate]      DATE  NOT NULL,
+    [Balance]        MONEY NOT NULL DEFAULT 0 CHECK ([Balance] >= 0),
+    [PendingBalance] MONEY NOT NULL DEFAULT 0 CHECK ([PendingBalance] >= 0),
+    [UserID]         INT   NOT NULL,
 
     FOREIGN KEY ([UserID]) REFERENCES [User] ([ID]),
 )
@@ -60,8 +60,8 @@ CREATE TABLE [Auction]
     [ProductID]    INT        NOT NULL,
     [Start]        DATETIME   NOT NULL,
     [End]          DATETIME   NOT NULL,
-    [StartPrice]   INT        NOT NULL CHECK ([StartPrice] >= 0),
-    [CurrentPrice] INT        NOT NULL CHECK ([CurrentPrice] >= 0),
+    [StartPrice]   MONEY      NOT NULL CHECK ([StartPrice] >= 0),
+    [CurrentPrice] MONEY      NOT NULL CHECK ([CurrentPrice] >= 0),
     [State]        VARCHAR(9) NOT NULL CHECK ([State] IN ('Open', 'Closed', 'Cancelled')) DEFAULT 'Open',
     [PublisherID]  INT        NOT NULL,
     [WinnerNIF]    INT        NULL,
@@ -76,7 +76,7 @@ CREATE TABLE [Bid]
     [ID]         INT      NOT NULL PRIMARY KEY IDENTITY,
     [AuctionID]  INT      NOT NULL,
     [BidderNIF]  INT      NOT NULL,
-    [Value]      INT      NOT NULL,
+    [Value]      MONEY    NOT NULL,
     [Date]       DATETIME DEFAULT GETDATE(),
 
     FOREIGN KEY ([AuctionID]) REFERENCES [Auction] ([ID]),
