+++ src/Data/Models/BidModel.cs
@@ -5,6 +5,6 @@ public class BidModel
     public int Id { get; }
     public int AuctionId { get; }
     public int BidderNif { get; }
-    public int Amount { get; }
+    public decimal Amount { get; }
     public DateTime Date { get; }
 }
